requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug
Pillow
pytesseract
openai